extend tutorial with inference
