+++ autorad/visualization/plotly_utils.py
@@ -6,7 +6,7 @@ import plotly.graph_objects as go
 from plotly.subplots import make_subplots
 from sklearn.metrics import precision_recall_curve, roc_auc_score, roc_curve
 
-from autorad.evaluation import eval_utils
+from autorad.utils import roc_utils
 
 from .matplotlib_utils import get_subplots_dimensions
 
@@ -130,7 +130,7 @@ def plot_waterfall(
     labels: Sequence[str] = ("positive", "negative"),
 ):
     if threshold is None:
-        threshold = eval_utils.get_youden_threshold(y_true, y_pred_proba)
+        threshold = roc_utils.get_youden_threshold(y_true, y_pred_proba)
 
     y_proba_rel_to_thr = [(val - threshold) for val in y_pred_proba]
     df = (
